+++ docs/en/changes/changes.md
@@ -116,6 +116,8 @@
 * Changed system variable `SW_SUPERDATASET_STORAGE_DAY_STEP` to `SW_STORAGE_ES_SUPER_DATASET_DAY_STEP` to be consistent with other ES storage related variables.
 * Fix ESEventQueryDAO missing metric_table boolQuery criteria.
 * Add default entity name(`_blank`) if absent to avoid NPE in the decoding. This caused `Can't split xxx id into 2 parts`.
+* Zipkin module support BanyanDB storage.
+* Zipkin traces query API, sort the result set by start time by default.
 
 #### UI
 
@@ -156,8 +158,12 @@
 * Add Python runtime metrics and cpu/memory utilization panels to General-Instance and Fass-Instance dashboards.
 * Enhance the legend of metrics graph widget with the summary table.
 * Add apache eventMesh logo file.
-* Fix conditions for trace profiling bug.
+* Fix conditions for trace profiling.
 * Fix tag keys list and duration condition.
+* Fix typo.
+* Fix condition logic for trace tree data.
+* Enhance tags component to search tags with the input value.
+* Fix topology loading style.
 
 #### Documentation